Add ability to "push" backup to a backup server
A lot had to change to add this capability since the script has
historically relied on the backup destination being local to the host
on which the script was running.

Passwordless ssh access to shell and rsync is required on the backup
server when doing push backups.... and this is inherently dangerous as
it grants your backup client way more privileges than it should have on
the backup server.